worldpose dataloader, plotting and profiler script

+++ scripts/inference.sh
@@ -3,6 +3,10 @@ CKPT_NAME=$1
 FILE_NAME=$2
 FPS=${3:-30}
 
+# Use system ffmpeg (has libx264) over conda ffmpeg (built with --disable-gpl)
+# Unset LD_LIBRARY_PATH so system ffmpeg doesn't load conda's shared libs
+FFMPEG="env -u LD_LIBRARY_PATH /usr/bin/ffmpeg"
+
 NAME="${FILE_NAME%.*}"
 EXT="${FILE_NAME##*.}"
 
@@ -15,7 +19,7 @@ mkdir -p $OUTPUT_PATH
 # convert video to frames
 case "$EXT" in
     mp4|avi|mov|mkv|flv|wmv|webm|mpeg|mpg)
-        ffmpeg -i ./demo/$FILE_NAME -f image2 -vf fps=${FPS}/1 -qscale 0 ${IMG_PATH}/%06d.jpg 
+        $FFMPEG -i ./demo/$FILE_NAME -f image2 -vf fps=${FPS}/1 -qscale 0 ${IMG_PATH}/%06d.jpg 
         ;;
     jpg|jpeg|png|bmp|gif|tiff|tif|webp|svg)
         cp ./demo/$FILE_NAME $IMG_PATH/000001.$EXT
@@ -30,17 +34,26 @@ END_COUNT=$(find "$IMG_PATH" -type f | wc -l)
 
 # inference with smplest_x
 PYTHONPATH=../:$PYTHONPATH \
+python main/inference_mod.py \
+    --num_gpus 1 \
+    --file_name ARG_CRO_220001 \
+    --ckpt_name smplest_x_h \
+    --start 1 \
+    --end 1238 \
+    --boxes_path ./demo/ARG_CRO_220001_boxes.npy
+
 python main/inference.py \
     --num_gpus 1 \
     --file_name $NAME \
     --ckpt_name $CKPT_NAME \
     --end $END_COUNT \
+    --multi_person \
 
 
 # convert frames to video
 case "$EXT" in
     mp4|avi|mov|mkv|flv|wmv|webm|mpeg|mpg)
-        ffmpeg -y -f image2 -r ${FPS} -i ${OUTPUT_PATH}/%06d.jpg -vcodec mjpeg -qscale 0 -pix_fmt yuv420p ./demo/result_${NAME}.mp4
+        $FFMPEG -y -f image2 -r ${FPS} -i ${OUTPUT_PATH}/%06d.jpg -vcodec libx264 -crf 18 -pix_fmt yuv420p ./demo/result_${NAME}.mp4
         ;;
     jpg|jpeg|png|bmp|gif|tiff|tif|webp|svg)
         cp $OUTPUT_PATH/000001.$EXT ./demo/result_$FILE_NAME
@@ -50,6 +63,10 @@ case "$EXT" in
         ;;
 esac
 
-rm -rf ./demo/input_frames
-rm -rf ./demo/output_frames
+if [ $? -eq 0 ]; then
+    rm -rf ./demo/input_frames
+    rm -rf ./demo/output_frames
+else
+    echo "ffmpeg failed. Output frames preserved at ${OUTPUT_PATH}"
+fi
